python/requirements.txt
paho-mqtt>=1.6.1
pygame>=2.5.2
python-dotenv>=1.0.0
